fix: move `tm2/pkg/keys/client` to `gno.land/pkg/keyscmd`
