+++ gno.land/pkg/keyscmd/add.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"context"
