+++ gno.land/pkg/keyscmd/broadcast.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"context"
