+++ gno.land/pkg/keyscmd/delete.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"context"
