+++ gno.land/pkg/keyscmd/delete_test.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"strings"
