+++ gno.land/pkg/keyscmd/export_test.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"fmt"
