+++ gno.land/pkg/keyscmd/generate_test.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"strings"
