+++ gno.land/pkg/keyscmd/import_test.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"fmt"
