+++ gno.land/pkg/keyscmd/list.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"context"
