+++ gno.land/pkg/keyscmd/list_test.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"testing"
