+++ gno.land/pkg/keyscmd/maketx.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"flag"
