+++ gno.land/pkg/keyscmd/query.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"context"
