+++ gno.land/pkg/keyscmd/run.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"context"
