+++ gno.land/pkg/keyscmd/send.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"context"
