+++ gno.land/pkg/keyscmd/sign.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"context"
