+++ gno.land/pkg/keyscmd/verify.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"context"
