+++ gno.land/pkg/keyscmd/verify_test.go
@@ -1,4 +1,4 @@
-package client
+package keyscmd
 
 import (
 	"encoding/hex"
